project_template/requirements.txt
Mezzanine
django-debug-toolbar
fabric
psycopg2
